use route resolvers to make sure lang is loaded before the subroute gets loaded
