+++ projects/bird-atlas/src/app/app.module.ts
@@ -19,7 +19,7 @@ import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { NgxDatatableModule } from '@achimha/ngx-datatable';
 import { API_BASE_URL } from 'projects/laji-api-client-b/src/laji-api-client-b.service';
 import { PlatformLocation } from '@angular/common';
-import { setLocale } from 'projects/laji/src/app/locale/locale.component';
+import { setLocale } from 'projects/laji/src/app/app-routing.modules';
 import { detectLangFromPath } from 'projects/laji/src/app/app.module';
 
 @NgModule({ exports: [],
