+++ projects/kerttu-global/src/app/locale/locale-es.component.ts
@@ -1,6 +1,4 @@
 import { Component } from '@angular/core';
-import { LocaleComponent } from 'projects/laji/src/app/locale/locale.component';
-import { PlatformService } from 'projects/laji/src/app/root/platform.service';
 import 'moment/locale/es';
 
 @Component({
@@ -8,12 +6,4 @@ import 'moment/locale/es';
     template: '<router-outlet></router-outlet>',
     standalone: false
 })
-export class LocaleEsComponent extends LocaleComponent {
-
-  constructor(
-    platformService: PlatformService,
-  ) {
-    super(platformService, 'es');
-  }
-
-}
+export class LocaleEsComponent {}
