+++ projects/kerttu-global/src/app/locale/locale-fr.component.ts
@@ -1,6 +1,4 @@
 import { Component } from '@angular/core';
-import { LocaleComponent } from 'projects/laji/src/app/locale/locale.component';
-import { PlatformService } from 'projects/laji/src/app/root/platform.service';
 import 'moment/locale/fr';
 
 @Component({
@@ -8,12 +6,4 @@ import 'moment/locale/fr';
     template: '<router-outlet></router-outlet>',
     standalone: false
 })
-export class LocaleFrComponent extends LocaleComponent {
-
-  constructor(
-    platformService: PlatformService,
-  ) {
-    super(platformService, 'fr');
-  }
-
-}
+export class LocaleFrComponent {}
