+++ projects/kerttu-global/src/app/locale/locale-zh.component.ts
@@ -1,6 +1,4 @@
 import { Component } from '@angular/core';
-import { LocaleComponent } from 'projects/laji/src/app/locale/locale.component';
-import { PlatformService } from 'projects/laji/src/app/root/platform.service';
 import 'moment/locale/zh-tw';
 
 @Component({
@@ -8,12 +6,4 @@ import 'moment/locale/zh-tw';
     template: '<router-outlet></router-outlet>',
     standalone: false
 })
-export class LocaleZhComponent extends LocaleComponent {
-
-  constructor(
-    platformService: PlatformService,
-  ) {
-    super(platformService, 'zh');
-  }
-
-}
+export class LocaleZhComponent {}
