+++ projects/laji/src/app/locale/locale-fi.component.ts
@@ -1,17 +1,8 @@
 import { Component } from '@angular/core';
-import { LocaleComponent } from './locale.component';
-import { PlatformService } from '../root/platform.service';
 
 @Component({
     selector: 'laji-locale-fi',
     template: '<router-outlet></router-outlet>',
     standalone: false
 })
-export class LocaleFiComponent extends LocaleComponent {
-
-  constructor(
-    platformService: PlatformService,
-  ) {
-    super(platformService, 'fi');
-  }
-}
+export class LocaleFiComponent {}
